Created first version of navbar component
